add speech recognition api and action files
