+++ client/actions/speech.js
@@ -1,0 +1,4 @@
+
+export function transcribeSpeechAction () {
+    console.log("speech action has been called")
+}
