+++ server/server.js
@@ -4,8 +4,9 @@ const cors = require('cors')
 const passport = require('passport')
 
 const authRoutes = require('./routes/auth')
-const meetingsRoutes = require('./routes/meetings')
 const usersRoutes = require('./routes/users')
+const speechRoutes = require('./routes/speech')
+// const meetingsRoutes = require('./routes/meetings')
 
 const server = express()
 
@@ -16,7 +17,8 @@ server.use(express.json())
 server.use(express.static(path.join(__dirname, '../public')))
 
 server.use('/api/auth', authRoutes)
-server.use('/api/meetings', meetingsRoutes)
 server.use('/api/users', usersRoutes)
+server.use('/api/speech', speechRoutes)
+// server.use('/api/meetings', meetingsRoutes)
 
 module.exports = server
