+++ webpack.config.js
@@ -19,5 +19,12 @@ module.exports = {
   resolve: {
     extensions: ['.js', '.jsx']
   },
+  // attempting to fix webpack vs. fs issue
+  target: 'node',
+  externals: {
+    fs: 'commonjs fs"'
+  },
+  node: {fs: "empty"},
+  //
   devtool: 'source-map'
 }
